scripts/requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
tqdm==4.66.1
pip==23.3.1
scipy==1.11.4
click==8.1.7
colorama~=0.4.4
setuptools==69.0.2
transforms==0.2.1
boto3==1.28.43
python-dotenv==1.0.0
botocore==1.31.43
google-cloud-texttospeech==2.15.0
openai==1.3.8
patreon~=0.5.0
fastapi==0.104.1
transformers==4.35.2
sentencepiece~=0.1.96
uvicorn==0.24.0.post1
pyrate-limiter~=3.1.0
torch~=2.0.1